+++ presto-main/src/main/java/com/facebook/presto/block/BlockEncodingManager.java
@@ -23,11 +23,11 @@ import com.facebook.presto.spi.block.FixedWidthBlockEncoding;
 import com.facebook.presto.spi.block.IntArrayBlockEncoding;
 import com.facebook.presto.spi.block.LazyBlockEncoding;
 import com.facebook.presto.spi.block.LongArrayBlockEncoding;
-import com.facebook.presto.spi.block.MapBlockEncoding;
+import com.facebook.presto.spi.block.Map.MapBlockEncoding;
+import com.facebook.presto.spi.block.Map.SingleMapBlockEncoding;
 import com.facebook.presto.spi.block.RowBlockEncoding;
 import com.facebook.presto.spi.block.RunLengthBlockEncoding;
 import com.facebook.presto.spi.block.ShortArrayBlockEncoding;
-import com.facebook.presto.spi.block.SingleMapBlockEncoding;
 import com.facebook.presto.spi.block.SingleRowBlockEncoding;
 import com.facebook.presto.spi.block.VariableWidthBlockEncoding;
 import com.facebook.presto.spi.type.TypeManager;
