+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/MapConstructor.java
@@ -24,8 +24,8 @@ import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
-import com.facebook.presto.spi.block.DuplicateMapKeyException;
-import com.facebook.presto.spi.block.MapBlockBuilder;
+import com.facebook.presto.spi.block.Map.DuplicateMapKeyException;
+import com.facebook.presto.spi.block.Map.MapBlockBuilder;
 import com.facebook.presto.spi.function.OperatorType;
 import com.facebook.presto.spi.type.MapType;
 import com.facebook.presto.spi.type.Type;
