+++ presto-main/src/test/java/com/facebook/presto/block/TestMapBlock.java
@@ -18,8 +18,8 @@ import com.facebook.presto.metadata.FunctionRegistry;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.block.ByteArrayBlock;
-import com.facebook.presto.spi.block.MapBlockBuilder;
-import com.facebook.presto.spi.block.SingleMapBlock;
+import com.facebook.presto.spi.block.Map.MapBlockBuilder;
+import com.facebook.presto.spi.block.Map.SingleMapBlock;
 import com.facebook.presto.spi.type.MapType;
 import com.facebook.presto.spi.type.TypeManager;
 import com.facebook.presto.sql.analyzer.FeaturesConfig;
