+++ presto-spi/src/main/java/com/facebook/presto/spi/block/BlockUtil.java
@@ -22,7 +22,7 @@ import static java.lang.Math.ceil;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 
-final class BlockUtil
+public final class BlockUtil
 {
     private static final double BLOCK_RESET_SKEW = 1.25;
 
@@ -34,7 +34,7 @@ final class BlockUtil
     {
     }
 
-    static void checkArrayRange(int[] array, int offset, int length)
+    public static void checkArrayRange(int[] array, int offset, int length)
     {
         requireNonNull(array, "array is null");
         if (offset < 0 || length < 0 || offset + length > array.length) {
@@ -42,21 +42,21 @@ final class BlockUtil
         }
     }
 
-    static void checkValidRegion(int positionCount, int positionOffset, int length)
+    public static void checkValidRegion(int positionCount, int positionOffset, int length)
     {
         if (positionOffset < 0 || length < 0 || positionOffset + length > positionCount) {
             throw new IndexOutOfBoundsException(format("Invalid position %s and length %s in block with %s positions", positionOffset, length, positionCount));
         }
     }
 
-    static void checkValidPosition(int position, int positionCount)
+    public static void checkValidPosition(int position, int positionCount)
     {
         if (position < 0 || position >= positionCount) {
             throw new IllegalArgumentException(format("Invalid position %s in block with %s positions", position, positionCount));
         }
     }
 
-    static int calculateNewArraySize(int currentSize)
+    public static int calculateNewArraySize(int currentSize)
     {
         // grow array by 50%
         long newSize = (long) currentSize + (currentSize >> 1);
@@ -74,7 +74,7 @@ final class BlockUtil
         return (int) newSize;
     }
 
-    static int calculateBlockResetSize(int currentSize)
+    public static int calculateBlockResetSize(int currentSize)
     {
         long newSize = (long) ceil(currentSize * BLOCK_RESET_SKEW);
 
@@ -88,7 +88,7 @@ final class BlockUtil
         return (int) newSize;
     }
 
-    static int calculateBlockResetBytes(int currentBytes)
+    public static int calculateBlockResetBytes(int currentBytes)
     {
         long newBytes = (long) ceil(currentBytes * BLOCK_RESET_SKEW);
         if (newBytes > MAX_ARRAY_SIZE) {
@@ -103,7 +103,7 @@ final class BlockUtil
      * with the first value set to 0.
      * If the range matches the entire <code>offsets</code> array,  the input array will be returned.
      */
-    static int[] compactOffsets(int[] offsets, int index, int length)
+    public static int[] compactOffsets(int[] offsets, int index, int length)
     {
         if (index == 0 && offsets.length == length + 1) {
             return offsets;
@@ -134,7 +134,7 @@ final class BlockUtil
      * If the range matches the entire array, the input array will be returned.
      * Otherwise, a copy will be returned.
      */
-    static boolean[] compactArray(boolean[] array, int index, int length)
+    public static boolean[] compactArray(boolean[] array, int index, int length)
     {
         if (index == 0 && length == array.length) {
             return array;
@@ -142,7 +142,7 @@ final class BlockUtil
         return Arrays.copyOfRange(array, index, index + length);
     }
 
-    static byte[] compactArray(byte[] array, int index, int length)
+    public static byte[] compactArray(byte[] array, int index, int length)
     {
         if (index == 0 && length == array.length) {
             return array;
@@ -150,7 +150,7 @@ final class BlockUtil
         return Arrays.copyOfRange(array, index, index + length);
     }
 
-    static short[] compactArray(short[] array, int index, int length)
+    public static short[] compactArray(short[] array, int index, int length)
     {
         if (index == 0 && length == array.length) {
             return array;
@@ -158,7 +158,7 @@ final class BlockUtil
         return Arrays.copyOfRange(array, index, index + length);
     }
 
-    static int[] compactArray(int[] array, int index, int length)
+    public static int[] compactArray(int[] array, int index, int length)
     {
         if (index == 0 && length == array.length) {
             return array;
@@ -166,7 +166,7 @@ final class BlockUtil
         return Arrays.copyOfRange(array, index, index + length);
     }
 
-    static long[] compactArray(long[] array, int index, int length)
+    public static long[] compactArray(long[] array, int index, int length)
     {
         if (index == 0 && length == array.length) {
             return array;
