+++ presto-spi/src/main/java/com/facebook/presto/spi/block/Map/ColumnarMap.java
@@ -11,7 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spi.block;
+package com.facebook.presto.spi.block.Map;
+
+import com.facebook.presto.spi.block.Block;
+import com.facebook.presto.spi.block.DictionaryBlock;
+import com.facebook.presto.spi.block.RunLengthEncodedBlock;
 
 import static java.util.Objects.requireNonNull;
