+++ presto-spi/src/main/java/com/facebook/presto/spi/block/Map/DuplicateMapKeyException.java
@@ -11,9 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spi.block;
+package com.facebook.presto.spi.block.Map;
 
 import com.facebook.presto.spi.ConnectorSession;
+import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.type.Type;
 
 import static java.lang.String.format;
