+++ presto-spi/src/main/java/com/facebook/presto/spi/block/Map/MapBlock.java
@@ -12,8 +12,9 @@
  * limitations under the License.
  */
 
-package com.facebook.presto.spi.block;
+package com.facebook.presto.spi.block.Map;
 
+import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.type.MapType;
 import com.facebook.presto.spi.type.Type;
 import org.openjdk.jol.info.ClassLayout;
@@ -25,7 +26,7 @@ import java.util.Arrays;
 import java.util.Optional;
 import java.util.function.BiConsumer;
 
-import static com.facebook.presto.spi.block.MapBlockBuilder.buildHashTable;
+import static com.facebook.presto.spi.block.Map.MapBlockBuilder.buildHashTable;
 import static io.airlift.slice.SizeOf.sizeOf;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
